Compatibility with HW2.2 distro
Packaging changes in 2.2 make querying the rpms very difficult, so
a script is added that queries hadoop commands to find the component
versions. Full version strings needed to find matching version in
maven repository.

This is only invoked on the HDP side, since proiblem is only
observed there, but the script should work as well for CDH.

Also change reg expressions looking for trx files to match both
old and newer distros.

Change-Id: Ia5963efdca98f56f0150cd9897dd0f4cffb43ded